# api/v1/openapi.v1.json
{
  "components": {
    "schemas": {
      "BackendCapabilitiesReport": {
        "additionalProperties": false,
        "properties": {
          "ann_status": {
            "title": "Ann Status",
            "type": "string"
          },
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "contracts": {
            "items": {
              "type": "string"
            },
            "title": "Contracts",
            "type": "array"
          },
          "deterministic_query": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deterministic Query"
          },
          "execution_modes": {
            "items": {
              "type": "string"
            },
            "title": "Execution Modes",
            "type": "array"
          },
          "isolation_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isolation Level"
          },
          "max_vector_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Vector Size"
          },
          "metrics": {
            "items": {
              "type": "string"
            },
            "title": "Metrics",
            "type": "array"
          },
          "replayable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replayable"
          },
          "storage_backends": {
            "items": {
              "$ref": "#/components/schemas/StorageBackendDescriptor"
            },
            "title": "Storage Backends",
            "type": "array"
          },
          "supports_ann": {
            "title": "Supports Ann",
            "type": "boolean"
          }
        },
        "required": [
          "backend",
          "contracts",
          "deterministic_query",
          "supports_ann",
          "replayable",
          "metrics",
          "max_vector_size",
          "isolation_level",
          "execution_modes",
          "ann_status",
          "storage_backends"
        ],
        "title": "BackendCapabilitiesReport",
        "type": "object"
      },
      "CreateRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateRequest",
        "type": "object"
      },
      "ExecutionArtifactRequest": {
        "additionalProperties": false,
        "properties": {
          "execution_contract": {
            "$ref": "#/components/schemas/ExecutionContract"
          }
        },
        "required": [
          "execution_contract"
        ],
        "title": "ExecutionArtifactRequest",
        "type": "object"
      },
      "ExecutionBudgetPayload": {
        "additionalProperties": false,
        "properties": {
          "max_error": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Error"
          },
          "max_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Latency Ms"
          },
          "max_memory_mb": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Memory Mb"
          }
        },
        "title": "ExecutionBudgetPayload",
        "type": "object"
      },
      "ExecutionContract": {
        "enum": [
          "deterministic",
          "non_deterministic"
        ],
        "title": "ExecutionContract",
        "type": "string"
      },
      "ExecutionIntent": {
        "enum": [
          "exact_validation",
          "reproducible_research",
          "exploratory_search",
          "production_retrieval"
        ],
        "title": "ExecutionIntent",
        "type": "string"
      },
      "ExecutionMode": {
        "enum": [
          "strict",
          "bounded",
          "exploratory"
        ],
        "title": "ExecutionMode",
        "type": "string"
      },
      "ExecutionRequestPayload": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "execution_budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionBudgetPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "execution_contract": {
            "$ref": "#/components/schemas/ExecutionContract"
          },
          "execution_intent": {
            "$ref": "#/components/schemas/ExecutionIntent"
          },
          "execution_mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "default": "strict"
          },
          "randomness_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RandomnessProfilePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Text"
          },
          "top_k": {
            "default": 5,
            "exclusiveMinimum": 0.0,
            "title": "Top K",
            "type": "integer"
          },
          "vector": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector"
          }
        },
        "required": [
          "execution_contract",
          "execution_intent"
        ],
        "title": "ExecutionRequestPayload",
        "type": "object"
      },
      "ExplainRequest": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "result_id": {
            "minLength": 1,
            "title": "Result Id",
            "type": "string"
          }
        },
        "required": [
          "result_id"
        ],
        "title": "ExplainRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestRequest": {
        "additionalProperties": false,
        "properties": {
          "documents": {
            "items": {
              "type": "string"
            },
            "title": "Documents",
            "type": "array"
          },
          "vectors": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Vectors",
            "type": "array"
          }
        },
        "required": [
          "documents",
          "vectors"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
      "RandomnessProfilePayload": {
        "additionalProperties": false,
        "properties": {
          "bounded": {
            "default": false,
            "title": "Bounded",
            "type": "boolean"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          }
        },
        "title": "RandomnessProfilePayload",
        "type": "object"
      },
      "StorageBackendDescriptor": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "persistence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persistence"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "title": "StorageBackendDescriptor",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "bijux-vex execution API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/artifact": {
      "post": {
        "operationId": "artifact_artifact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionArtifactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Artifact Artifact Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifact"
      }
    },
    "/capabilities": {
      "get": {
        "operationId": "capabilities_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackendCapabilitiesReport"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Capabilities"
      }
    },
    "/create": {
      "post": {
        "operationId": "create_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Create Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create"
      }
    },
    "/execute": {
      "post": {
        "operationId": "execute_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Execute Execute Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute"
      }
    },
    "/explain": {
      "post": {
        "operationId": "explain_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Explain Explain Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain"
      }
    },
    "/ingest": {
      "post": {
        "operationId": "ingest_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ingest Ingest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest"
      }
    },
    "/replay": {
      "post": {
        "operationId": "replay_replay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Replay Replay Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay"
      }
    }
  }
}